+++ Sources/CodexBarCore/AutoreleasePoolCompat.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-#if os(Linux)
+#if os(Linux) || os(Windows)
 @discardableResult
 func autoreleasepool<Result>(_ work: () throws -> Result) rethrows -> Result {
     try work()