+++ Tests/CodexBarTests/CostUsageStoreTests.swift
@@ -1013,6 +1013,7 @@ extension CostUsageStoreTests {
             "2d17f4981b78d07f",
             "99848ef16ca7e069",
             "ad52067daa4e07eb",
+            "4563a0078e5bfa32",
         ])
         let predecessorHash = "2d17f4981b78d07f"
         let predecessorVersion = CostUsageStore.combinedSchemaVersion(